fix: infinite auth/me fetch
